Narrowing scope.
Attempting a total backend and frontend rewrite is suicide. (Well,
maybe not for Toph.) Bongo from here on out is going to be a CMS
and an API. Then we'll hook the existing frontend up to the API.
Then we'll think about writing a nice new javascript frontend.
What should we call it? Bonusbone? Angularus? BonJS?